Fixed failing Login query
