+++ nbproject/private/private.xml
@@ -2,11 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/Users/rwothoromo/apps/general/TGS/src/database/DeleteDatabase.java</file>
-            <file>file:/Users/rwothoromo/apps/general/TGS/src/database/Connect.java</file>
-            <file>file:/Users/rwothoromo/apps/general/TGS/.gitignore</file>
-            <file>file:/Users/rwothoromo/apps/general/TGS/nbproject/project.properties</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
