+++ nbproject/project.properties
@@ -54,6 +54,7 @@ javac.test.classpath=\
     ${build.classes.dir}
 javac.test.modulepath=\
     ${javac.modulepath}
+javadoc.additionalparam=
 javadoc.author=false
 javadoc.html5=false
 javadoc.noindex=false
@@ -65,7 +66,7 @@ javadoc.use=false
 javadoc.version=false
 jlink.launcher=false
 jlink.launcher.name=TGS
-main.class=user.Login
+main.class=auth.Login
 mkdist.disabled=false
 platform.active=default_platform
 run.classpath=\
