+++ src/admin/EditPassword.java
@@ -1,20 +1,21 @@
-package user;
+package admin;
 
 import database.Connect;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import javax.swing.JOptionPane;
+import auth.Login;
 
-public class SUChangePassword extends javax.swing.JFrame {
-    Connection x = null;
-    public SUChangePassword() {
+public class EditPassword extends javax.swing.JFrame {
+    Connection conn = null;
+    public EditPassword() {
         initComponents();
-        x = Connect.dbConnector();
+        conn = Connect.dbConnector();
     }
 
     /**
-     * This method is called from within the constructor to initialize the form.
+     * This method is called FROM within the constructor to initialize the form.
      * WARNING: Do NOT modify this code. The content of this method is always
      * regenerated by the Form Editor.
      */
@@ -185,12 +186,12 @@ public class SUChangePassword extends javax.swing.JFrame {
         String OldPassword = txtOldPassword.getText();
         String NewPassword = txtNewPassword.getText();
         String ConfirmPassword = txtConfirmPassword.getText();
-        String sqlQueryPwd = "select Password from super_user where IDNO=? and Password=?";
-        String sqlQueryP = "update super_user set Password='"
-                + ConfirmPassword + "' where IDNO='" + IDNO + "'";
+        String sqlQueryPwd = "SELECT Password FROM super_user WHERE IDNO=? AND Password=?";
+        String sqlQueryP = "UPDATE super_user SET Password='"
+                + ConfirmPassword + "' WHERE IDNO='" + IDNO + "'";
         try {
-            PreparedStatement ps = x.prepareStatement(sqlQueryPwd);
-            PreparedStatement ps1 = x.prepareStatement(sqlQueryP);
+            PreparedStatement ps = conn.prepareStatement(sqlQueryPwd);
+            PreparedStatement ps1 = conn.prepareStatement(sqlQueryP);
             ps.setString(1, IDNO);
             ps.setString(2, OldPassword);
             ResultSet rs = ps.executeQuery();
@@ -216,9 +217,9 @@ public class SUChangePassword extends javax.swing.JFrame {
      * @param args the command line arguments
      */
     public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
+        /* Set the Nimbus look AND feel */
+        //<editor-fold defaultstate="collapsed" desc=" Look AND feel setting code (optional) ">
+        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look AND feel.
          * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
          */
         try {
@@ -229,23 +230,27 @@ public class SUChangePassword extends javax.swing.JFrame {
                 }
             }
         } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(SUChangePassword.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(EditPassword.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(SUChangePassword.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(EditPassword.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(SUChangePassword.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(EditPassword.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(SUChangePassword.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
+            java.util.logging.Logger.getLogger(EditPassword.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
         }
         //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
+        //</editor-fold>
 
-        /* Create and display the form */
+        /* Create AND display the form */
         java.awt.EventQueue.invokeLater(new Runnable() {
             public void run() {
-                new SUChangePassword().setVisible(true);
+                new EditPassword().setVisible(true);
             }
         });
     }
+    
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton Back;
     private javax.swing.JButton btnSuperUserSubmit;
