+++ src/constants/Constants.java
@@ -1,0 +1,16 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package constants;
+
+/**
+ *
+ * @author rwothoromo
+ */
+public class Constants {
+    public static final String DB_PATH = System.getenv("DB_PATH");
+    public static final String DB_USER = System.getenv("DB_USER");
+    public static final String DB_PASSWORD = System.getenv("DB_PASSWORD");
+}
